refactor: delete @sidecar/realtime, keep the phone's mint document
The phone's Realtime mint document moves to @sidecar/actions beside the
action tools it carries, the credential reader joins @sidecar/hosted's
realtime-contract, and the package, its doors, its required docs, and the
knip ignore go with it.
