+++ apps/web/package.json
@@ -30,7 +30,6 @@
     "@sidecar/brain": "workspace:*",
     "@sidecar/hosted": "workspace:*",
     "@sidecar/panel": "workspace:*",
-    "@sidecar/realtime": "workspace:*",
     "@sidecar/session": "workspace:*",
     "@sidecar/settings": "workspace:*",
     "@sidecar/surface": "workspace:*",
