+++ apps/web/server/core.ts
@@ -43,7 +43,6 @@ export * from "../../../packages/analytics/src/index.js";
 export * from "../../../packages/brain/src/index.js";
 export * from "../../../packages/brain/src/store/shapes.js";
 export * from "../../../packages/hosted/src/index.js";
-export * from "../../../packages/realtime/src/index.js";
 export * from "../../../packages/runtime/src/vocabulary.js";
 export * from "../../../packages/session/src/index.js";
 export * from "../../../packages/session/src/ui-messages/index.js";
