+++ packages/actions/src/index.ts
@@ -6,3 +6,4 @@ export * from "./actions.js";
 export * from "./adapter-requests.js";
 export * from "./admit.js";
 export * from "./memory.js";
+export * from "./remote-mint-legacy.js";
