+++ packages/devtrace/package.json
@@ -14,7 +14,6 @@
   },
   "dependencies": {
     "@sidecar/brain": "workspace:*",
-    "@sidecar/realtime": "workspace:*",
     "@sidecar/runtime": "workspace:*",
     "@sidecar/wire": "workspace:*"
   },
