+++ packages/hosted/src/realtime-contract.ts
@@ -1,3 +1,5 @@
+import { isRecord, text, type UnparsedWireValue, wholeNumber } from "@sidecar/wire";
+
 /** The OpenAI path a Realtime connection opens after minting. */
 export const REALTIME_CALLS_PATH = "/realtime/calls";
 
@@ -23,4 +25,33 @@ export interface RealtimeConnection extends RealtimeCredential {
 
 export function realtimeCredentialIsUsable(credential: RealtimeCredential, now: number): boolean {
   return credential.expiresAt > now;
+}
+
+/**
+ * Reads an untrusted mint response into the contract. Anything that does not
+ * carry a usable secret and expiry is discarded rather than repaired, so a
+ * malformed response leaves voice unavailable instead of half-configured. A
+ * response that omits its model is labelled with the one it was minted for,
+ * which the caller knows and this package does not.
+ */
+export function realtimeCredentialFromResponse(
+  payload: UnparsedWireValue,
+  mintedModel: string,
+): RealtimeCredential | undefined {
+  if (!isRecord(payload)) return undefined;
+
+  const value = text(payload.value);
+  if (!value) return undefined;
+
+  const expiresAtSeconds = wholeNumber(payload.expires_at);
+  if (expiresAtSeconds === undefined || expiresAtSeconds <= 0) return undefined;
+
+  const session = isRecord(payload.session) ? payload.session : undefined;
+  const model = session ? text(session.model) : undefined;
+
+  return {
+    value,
+    expiresAt: Math.floor(expiresAtSeconds * 1000),
+    model: model ?? mintedModel,
+  };
 }
