--- packages/realtime/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@sidecar/realtime",
-  "version": "0.5.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.js",
-    "./speech": "./src/speech-wire.js"
-  },
-  "types": "./src/index.ts",
-  "scripts": {
-    "test": "tsx --test 'src/**/*.test.ts'",
-    "typecheck": "tsc --project tsconfig.json"
-  },
-  "dependencies": {
-    "@sidecar/actions": "workspace:*",
-    "@sidecar/guide": "workspace:*",
-    "@sidecar/hosted": "workspace:*",
-    "@sidecar/session": "workspace:*",
-    "@sidecar/wire": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/node": "catalog:",
-    "tsx": "catalog:",
-    "typescript": "catalog:"
-  }
-}
