--- packages/realtime/src/index.ts
@@ -1,75 +1,0 @@
-export { conversationSeedEvents } from "./conversation-seed.js";
-export {
-  MAXIMUM_PRESS_AUDIO_MS,
-  PRESS_AUDIO_SAMPLE_RATE,
-  PressAudioBuffer,
-} from "./press-audio.js";
-export {
-  ARRIVAL_SPEECH_KIND,
-  type ArrivalSpeech,
-  arrivalSpeechEvents,
-  BRIEFING_SPEECH_KIND,
-  type BriefingSpeech,
-  briefingSpeechEvents,
-  CALENDAR_ONBOARDING_SPEECH_KIND,
-  type CalendarOnboardingSpeech,
-  isProactiveSpeechTurn,
-  type ProactiveSpeechTurn,
-} from "./proactive-speech.js";
-export {
-  introductionSessionConfig,
-  REALTIME_CLIENT_SECRETS_PATH,
-  REALTIME_MINT_OUTCOME,
-  REALTIME_TRUNCATION,
-  type RealtimeDiagnostics,
-  type RealtimeMintOutcome,
-  type RealtimeSessionOptions,
-  realtimeClientSecretRequest,
-  realtimeCredentialFromResponse,
-  realtimeMintExplanation,
-  realtimeSessionConfig,
-  remoteRealtimeClientSecretRequest,
-} from "./realtime-credentials.js";
-export {
-  cancelResponseEvents,
-  clearInputAudioEvents,
-  clearOutputAudioEvents,
-  decodeRealtimePayload,
-  functionCallFollowUpEvents,
-  inputAudioAppendEvents,
-  inputAudioFormatUpdateEvents,
-  outputSpeedUpdateEvents,
-  type ParsedRealtimeFunctionCall,
-  type ParsedRealtimeServerEvent,
-  parseRealtimeServerEvent,
-  pushToTalkCommitEvents,
-  REALTIME_CLIENT_EVENT,
-  REALTIME_SERVER_EVENT,
-  REALTIME_STATUS,
-  type RealtimeStatus,
-  truncateResponseEvents,
-  voiceExchangeActive,
-} from "./realtime-events.js";
-export {
-  ASK_BRAIN_TOOL,
-  type MouthToolDefinition,
-  mouthToolDefinitions,
-} from "./realtime-instructions.js";
-export {
-  isRealtimeVoice,
-  isRealtimeVoiceSpeed,
-  REALTIME_DEFAULTS,
-  REALTIME_VOICE,
-  REALTIME_VOICE_LIST,
-  REALTIME_VOICE_SPEED,
-  REALTIME_VOICE_SPEED_LIST,
-  type RealtimeVoice,
-  type RealtimeVoiceSpeed,
-} from "./realtime-voice-settings.js";
-export {
-  BRIEFING_INPUT_MARKER,
-  NOTE_MARKER,
-  responseTurn,
-  SCENE,
-  sessionInstructions,
-} from "./voice-scene.js";
