--- packages/realtime/src/realtime-events.ts
@@ -1,607 +1,0 @@
-import type { RealtimeFunctionCall } from "@sidecar/actions";
-import {
-  isRecord,
-  isWireString,
-  text,
-  type UnparsedWireValue,
-  type WireRecord,
-} from "@sidecar/wire";
-import { PRESS_AUDIO_SAMPLE_RATE } from "./press-audio.js";
-import { trimmedText } from "./trimmed-text.js";
-
-/**
- * The Realtime wire grammar: how far a call has progressed, the events both
- * sides send, the outbound builders that speak them, and the parser that
- * reads inbound events so a second file cannot re-encode the grammar.
- */
-
-/** The Realtime session shape a client secret is minted against. */
-export const REALTIME_SESSION_TYPE = "realtime";
-
-/** How far the voice loop has progressed, as the main process and UI both read it. */
-export const REALTIME_STATUS = {
-  /** No credentials are configured, so the voice experience is off. */
-  UNAVAILABLE: "unavailable",
-  IDLE: "idle",
-  CONNECTING: "connecting",
-  /** Connected with the microphone held closed until push-to-talk is pressed. */
-  READY: "ready",
-  LISTENING: "listening",
-  RESPONDING: "responding",
-  FAILED: "failed",
-} as const;
-
-export type RealtimeStatus = (typeof REALTIME_STATUS)[keyof typeof REALTIME_STATUS];
-
-/**
- * The exchange is live from the press to the end of the reply — the call
- * coming up, a turn being held, Luke speaking — and the media duck follows it.
- * It is also the whole of the session's turn arbitration: the Realtime API
- * answers one turn at a time, so while this stands nothing may open another,
- * and a caller told it was refused decides what to show instead.
- */
-export function voiceExchangeActive(status: RealtimeStatus): boolean {
-  return (
-    status === REALTIME_STATUS.CONNECTING ||
-    status === REALTIME_STATUS.LISTENING ||
-    status === REALTIME_STATUS.RESPONDING
-  );
-}
-
-export const REALTIME_CLIENT_EVENT = {
-  SESSION_UPDATE: "session.update",
-  /**
-   * Adds audio to the input buffer over the data channel. On a WebRTC call the
-   * microphone track already feeds that buffer, so this exists for the audio
-   * the track cannot carry: the words spoken while the call was still
-   * connecting, captured locally and flushed once the channel opens.
-   */
-  INPUT_AUDIO_BUFFER_APPEND: "input_audio_buffer.append",
-  INPUT_AUDIO_BUFFER_COMMIT: "input_audio_buffer.commit",
-  INPUT_AUDIO_BUFFER_CLEAR: "input_audio_buffer.clear",
-  CONVERSATION_ITEM_CREATE: "conversation.item.create",
-  /**
-   * Removes an item from the conversation. Only ever used on context Luke
-   * himself put there — a roster, a guide, a projects list, an issue board —
-   * and only to make room for the same context said again. Nothing the
-   * developer said is ever deleted.
-   */
-  CONVERSATION_ITEM_DELETE: "conversation.item.delete",
-  RESPONSE_CREATE: "response.create",
-  RESPONSE_CANCEL: "response.cancel",
-  /**
-   * WebRTC only, and the only event that actually stops a reply being heard.
-   * The model generates faster than it speaks, so by the time someone talks
-   * over Luke the rest of his sentence has already been sent — cancelling stops
-   * him producing more and does nothing about what is already on the way.
-   */
-  OUTPUT_AUDIO_BUFFER_CLEAR: "output_audio_buffer.clear",
-  /**
-   * Trims an assistant message to what was actually heard. Without it the model
-   * carries on believing it said the whole reply, so it can answer a follow-up
-   * by referring back to a sentence that was cut off before it was spoken.
-   */
-  CONVERSATION_ITEM_TRUNCATE: "conversation.item.truncate",
-} as const;
-
-export const REALTIME_SERVER_EVENT = {
-  RESPONSE_CREATED: "response.created",
-  /** The server confirming it dropped the audio it had queued for us. */
-  OUTPUT_AUDIO_BUFFER_CLEARED: "output_audio_buffer.cleared",
-  /** Names the message a reply is being spoken into, which is what a truncate cuts. */
-  RESPONSE_OUTPUT_ITEM_ADDED: "response.output_item.added",
-  /**
-   * Luke has stopped speaking — the server's own word for it, sent once the
-   * audio it queued has drained. WebRTC only, and absent from the API reference
-   * while being what every WebRTC client needs, so the silence heuristic stays
-   * behind it as a backstop rather than being replaced outright.
-   */
-  OUTPUT_AUDIO_BUFFER_STOPPED: "output_audio_buffer.stopped",
-  /**
-   * Luke is audible again — the server streaming into a buffer that had
-   * drained. WebRTC only, like its stop, and what makes a mid-reply stop
-   * readable at all: a reply with more than one thing to say can drain the
-   * buffer between them, and only this event says that drain was a pause
-   * rather than the ending.
-   */
-  OUTPUT_AUDIO_BUFFER_STARTED: "output_audio_buffer.started",
-  /**
-   * The words of the reply, arriving as they are generated. They run ahead of
-   * the audio — the model produces text faster than it speaks it — so they
-   * caption the reply in progress rather than subtitling word by word.
-   */
-  RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DELTA: "response.output_audio_transcript.delta",
-  /** The reply's complete text, which supersedes whatever the deltas built. */
-  RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DONE: "response.output_audio_transcript.done",
-  /** The server has made one developer audio turn into a conversation item. */
-  INPUT_AUDIO_BUFFER_COMMITTED: "input_audio_buffer.committed",
-  /**
-   * The developer's own spoken words, arriving as the service transcribes
-   * them. They only preview the completed transcript below — the settled
-   * words supersede whatever the deltas built — so a surface may show the
-   * ask taking shape while the history still records only what completed.
-   */
-  INPUT_AUDIO_TRANSCRIPTION_DELTA: "conversation.item.input_audio_transcription.delta",
-  /**
-   * The developer's own spoken turn, as the service transcribed it. It
-   * arrives on its own clock — transcription runs beside the reply, not ahead
-   * of it — and it is the one way their spoken words settle as text at all,
-   * so the conversation history can hold both halves of the exchange.
-   */
-  INPUT_AUDIO_TRANSCRIPTION_COMPLETED: "conversation.item.input_audio_transcription.completed",
-  /**
-   * A spoken turn whose transcription the service gave up on. No words ever
-   * arrive for it, so whatever preview its deltas built must leave rather
-   * than stand forever for a completion that is not coming.
-   */
-  INPUT_AUDIO_TRANSCRIPTION_FAILED: "conversation.item.input_audio_transcription.failed",
-  RESPONSE_DONE: "response.done",
-  ERROR: "error",
-} as const;
-
-/** Clears audio already queued for playback, naming the request for error correlation. */
-export function clearOutputAudioEvents(eventId: string): readonly WireRecord[] {
-  const clearEventId = trimmedText(eventId);
-  if (!clearEventId) return [];
-  return [{ type: REALTIME_CLIENT_EVENT.OUTPUT_AUDIO_BUFFER_CLEAR, event_id: clearEventId }];
-}
-
-/**
- * Builds the events that stop a reply the developer is talking over.
- *
- * Both are needed and in this order. Cancelling stops the model producing more
- * of the reply; it says nothing about the audio it already produced, which the
- * server has sent ahead because it generates faster than speech. Clearing the
- * output buffer is what drops that, and doing it second means the server is not
- * still filling the buffer as it empties it.
- */
-export function cancelResponseEvents(input: {
-  cancellationEventId: string;
-  clearEventId: string;
-}): readonly WireRecord[] {
-  const cancellationEventId = trimmedText(input.cancellationEventId);
-  const clearEvents = clearOutputAudioEvents(input.clearEventId);
-  if (!cancellationEventId || clearEvents.length === 0) return [];
-  return [
-    { type: REALTIME_CLIENT_EVENT.RESPONSE_CANCEL, event_id: cancellationEventId },
-    ...clearEvents,
-  ];
-}
-
-/**
- * Builds the event that trims a cut-off reply to what was heard of it.
- *
- * Stopping the sound and correcting the record are two different things. The
- * first is what the developer notices; without the second the model believes it
- * said every word it generated, and will happily refer back to a sentence that
- * never reached the room.
- *
- * `audioEndMs` is how long the reply was audible, measured on a wall clock —
- * which can outrun the audio itself when playback stalls, or when the stop
- * lands at the reply's very end. The server answers such a trim with a refusal
- * and then clamps it to the audio's real end and truncates anyway, so the
- * record is right either way; the refusal names no event of its own on the
- * wire, and the session recognizes it by its sentence instead.
- */
-export function truncateResponseEvents(input: {
-  itemId: string;
-  audioEndMs: number;
-  truncationEventId: string;
-}): readonly WireRecord[] {
-  const truncationEventId = trimmedText(input.truncationEventId);
-  if (
-    !trimmedText(input.itemId) ||
-    !truncationEventId ||
-    !Number.isFinite(input.audioEndMs) ||
-    input.audioEndMs <= 0
-  ) {
-    return [];
-  }
-  return [
-    {
-      type: REALTIME_CLIENT_EVENT.CONVERSATION_ITEM_TRUNCATE,
-      event_id: truncationEventId,
-      item_id: input.itemId,
-      // One audio part per assistant message, so the first is the reply.
-      content_index: 0,
-      audio_end_ms: Math.floor(input.audioEndMs),
-    },
-  ];
-}
-
-const BASE64_ALPHABET = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-
-/**
- * Encodes bytes without reaching for an environment: `btoa` wants a binary
- * string and Node's `Buffer` does not exist in a sandboxed renderer, and the
- * audio event is this module's grammar, so its encoding lives here too.
- */
-function base64FromBytes(bytes: Uint8Array): string {
-  let encoded = "";
-  for (let index = 0; index < bytes.length; index += 3) {
-    const first = bytes[index] ?? 0;
-    const second = bytes[index + 1];
-    const third = bytes[index + 2];
-    encoded += BASE64_ALPHABET[first >> 2];
-    encoded += BASE64_ALPHABET[((first & 0b11) << 4) | ((second ?? 0) >> 4)];
-    encoded +=
-      second === undefined ? "=" : BASE64_ALPHABET[((second & 0b1111) << 2) | ((third ?? 0) >> 6)];
-    encoded += third === undefined ? "=" : BASE64_ALPHABET[third & 0b111111];
-  }
-  return encoded;
-}
-
-/**
- * Builds the event that declares what the appended audio is, ahead of
- * appending any.
- *
- * On a WebRTC call the negotiated input is Opus at the codec's own rate, so
- * nothing about the call itself says how base64 audio arriving over the data
- * channel should be read — and 24kHz PCM read at any other rate is not heard
- * wrong, it is not heard at all. The keyed mint pins the same format, but the
- * hosted mint composes its session on the service, so the one place every
- * call can be pinned from is the channel itself: this travels at the start of
- * every captured turn, before the clear and the appends it speaks for.
- */
-export function inputAudioFormatUpdateEvents(): readonly WireRecord[] {
-  return [
-    {
-      type: REALTIME_CLIENT_EVENT.SESSION_UPDATE,
-      session: {
-        type: REALTIME_SESSION_TYPE,
-        audio: { input: { format: { type: "audio/pcm", rate: PRESS_AUDIO_SAMPLE_RATE } } },
-      },
-    },
-  ];
-}
-
-/**
- * Builds the event that adds one captured chunk to the input audio buffer.
- *
- * The samples travel as the format {@link inputAudioFormatUpdateEvents} pins —
- * 16-bit PCM at the press capture's own rate, little-endian whatever this
- * machine is, base64 over the data channel. An empty chunk builds nothing:
- * there is no audio to add, and the API refuses an empty append rather than
- * ignoring it.
- */
-export function inputAudioAppendEvents(audio: Int16Array): readonly WireRecord[] {
-  if (audio.length === 0) return [];
-  const bytes = new Uint8Array(audio.length * 2);
-  for (let index = 0; index < audio.length; index += 1) {
-    const sample = audio[index] ?? 0;
-    bytes[index * 2] = sample & 0xff;
-    bytes[index * 2 + 1] = (sample >> 8) & 0xff;
-  }
-  return [
-    {
-      type: REALTIME_CLIENT_EVENT.INPUT_AUDIO_BUFFER_APPEND,
-      audio: base64FromBytes(bytes),
-    },
-  ];
-}
-
-/** Builds the events that close a push-to-talk turn and ask for a reply. */
-export function pushToTalkCommitEvents(): readonly WireRecord[] {
-  return [
-    { type: REALTIME_CLIENT_EVENT.INPUT_AUDIO_BUFFER_COMMIT },
-    { type: REALTIME_CLIENT_EVENT.RESPONSE_CREATE },
-  ];
-}
-
-/**
- * Builds the event that empties the input audio buffer.
- *
- * A turn both opens and is abandoned with this. With turn detection disabled
- * the server keeps every byte received since the last commit, and a muted track
- * still transmits, so a turn that did not start from an empty buffer would
- * commit however long the call had been sitting idle along with what was said.
- */
-export function clearInputAudioEvents(): readonly WireRecord[] {
-  return [{ type: REALTIME_CLIENT_EVENT.INPUT_AUDIO_BUFFER_CLEAR }];
-}
-
-/**
- * Builds the event that changes how fast a live call speaks, from its next
- * reply on.
- *
- * The pace is the one output setting the API lets a running session change —
- * a session's voice locks the moment the model first speaks, so a changed
- * voice is heard by opening a new call rather than by any event this module
- * could build. The API applies a pace only between turns, which the caller is
- * left to time; a pace that is not a usable number builds nothing rather than
- * an update the API would refuse.
- */
-export function outputSpeedUpdateEvents(speed: number): readonly WireRecord[] {
-  if (!Number.isFinite(speed) || speed <= 0) return [];
-  return [
-    {
-      type: REALTIME_CLIENT_EVENT.SESSION_UPDATE,
-      session: { type: REALTIME_SESSION_TYPE, audio: { output: { speed } } },
-    },
-  ];
-}
-
-/**
- * One tool call the model made, as it arrives inside a finished response: the
- * action's own name and arguments, and the id its answer has to carry back. The
- * id stops here — a validator is handed the call itself, which is why
- * `RealtimeFunctionCall` in `@sidecar/actions` has no room for one.
- */
-export interface ParsedRealtimeFunctionCall extends RealtimeFunctionCall {
-  callId: string;
-}
-
-/**
- * An inbound Realtime event the conversation acts on. The wire names stay the
- * discriminant so a switch is a switch on the protocol, not a second vocabulary.
- */
-export type ParsedRealtimeServerEvent =
-  | { type: typeof REALTIME_SERVER_EVENT.RESPONSE_CREATED; responseId?: string }
-  | { type: typeof REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_ITEM_ADDED; itemId?: string }
-  | {
-      type: typeof REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DELTA;
-      itemId?: string;
-      delta: string;
-    }
-  | {
-      type: typeof REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DONE;
-      itemId?: string;
-      transcript: string;
-    }
-  | {
-      type: typeof REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_DELTA;
-      itemId: string;
-      delta: string;
-    }
-  | {
-      type: typeof REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_COMPLETED;
-      itemId: string;
-      transcript: string;
-    }
-  | { type: typeof REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_FAILED; itemId: string }
-  | { type: typeof REALTIME_SERVER_EVENT.INPUT_AUDIO_BUFFER_COMMITTED; itemId: string }
-  | { type: typeof REALTIME_SERVER_EVENT.OUTPUT_AUDIO_BUFFER_STOPPED; responseId?: string }
-  | { type: typeof REALTIME_SERVER_EVENT.OUTPUT_AUDIO_BUFFER_STARTED; responseId?: string }
-  | {
-      type: typeof REALTIME_SERVER_EVENT.RESPONSE_DONE;
-      responseId?: string;
-      calls: readonly ParsedRealtimeFunctionCall[];
-      /**
-       * Whether the finished response made any sound, absent when the payload
-       * carried no output to read it from. A reply of pure silence — a
-       * success answered without a word — has nothing to play out, so its
-       * turn may end here instead of waiting on quiet that never comes.
-       */
-      hasAudio?: boolean;
-    }
-  /**
-   * `eventId` names the client event the service is complaining about, when it
-   * says. It is what lets a caller tell an error meant for the developer from
-   * the answer to something it asked for itself.
-   */
-  | {
-      type: typeof REALTIME_SERVER_EVENT.ERROR;
-      message: string;
-      eventId?: string;
-      errorType?: string;
-      errorCode?: string;
-    };
-
-function optionalString(value: UnparsedWireValue): string | undefined {
-  return text(value);
-}
-
-function recordField(record: WireRecord, key: string): WireRecord | undefined {
-  const value = record[key];
-  return isRecord(value) ? value : undefined;
-}
-
-/**
- * The tool calls a `response.done` event carries, if any. Read from the
- * finished response rather than streamed deltas: a call is acted on whole or
- * not at all, and the finished response is the only place it is whole.
- */
-function functionCallsFromDone(event: WireRecord): readonly ParsedRealtimeFunctionCall[] {
-  const response = recordField(event, "response");
-  const output = Array.isArray(response?.output) ? response.output : [];
-  return output.filter(isRecord).flatMap((item) => {
-    if (item.type !== "function_call") return [];
-    const name = text(item.name) ?? "";
-    const callId = text(item.call_id) ?? "";
-    const argumentsJson = text(item.arguments) ?? "";
-    return name && callId ? [{ name, callId, argumentsJson }] : [];
-  });
-}
-
-/**
- * Whether a `response.done` event's response produced any audio, read from
- * its own output items. Unknown when there is no output array to read,
- * and unknown must not pass for silent: only a response that says what it
- * made may say it made no sound.
- */
-function audioFromDone(event: WireRecord): boolean | undefined {
-  const response = recordField(event, "response");
-  const output = response && Array.isArray(response.output) ? response.output : undefined;
-  if (!output) return undefined;
-  return output.filter(isRecord).some((item) => {
-    const content = Array.isArray(item.content) ? item.content : [];
-    return content.filter(isRecord).some((part) => part.type === "output_audio");
-  });
-}
-
-/**
- * Decodes one data-channel payload to the record it carries, or nothing. This
- * is the wire format's own reader — exported so a tap on the channel reads
- * the payload the same way the parser below does, rather than re-encoding
- * the grammar in a second style.
- */
-export function decodeRealtimePayload(data: UnparsedWireValue): WireRecord | undefined {
-  let payload: UnparsedWireValue = data;
-  if (isWireString(data)) {
-    try {
-      // SAFETY: JSON.parse returns a runtime value; isRecord validates the object contract.
-      payload = JSON.parse(data) as UnparsedWireValue;
-    } catch (error) {
-      if (error instanceof SyntaxError) return undefined;
-      throw error;
-    }
-  }
-  return isRecord(payload) ? payload : undefined;
-}
-
-/**
- * Reads one inbound Realtime event. The wire format is this module's: a JSON
- * string from the data channel, or an already-decoded payload, is accepted so
- * a second file cannot re-encode the grammar in a second style. Anything that
- * is not one of the events the conversation acts on is discarded rather than
- * repaired.
- */
-export function parseRealtimeServerEvent(
-  data: UnparsedWireValue,
-): ParsedRealtimeServerEvent | undefined {
-  const event = decodeRealtimePayload(data);
-  if (!event) return undefined;
-
-  switch (event.type) {
-    case REALTIME_SERVER_EVENT.RESPONSE_CREATED: {
-      const responseId = optionalString(recordField(event, "response")?.id);
-      const parsed: ParsedRealtimeServerEvent = {
-        type: REALTIME_SERVER_EVENT.RESPONSE_CREATED,
-      };
-      if (responseId) parsed.responseId = responseId;
-      return parsed;
-    }
-    case REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_ITEM_ADDED: {
-      const itemId = optionalString(recordField(event, "item")?.id);
-      const parsed: ParsedRealtimeServerEvent = {
-        type: REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_ITEM_ADDED,
-      };
-      if (itemId) parsed.itemId = itemId;
-      return parsed;
-    }
-    case REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DELTA: {
-      if (!isWireString(event.delta) || event.delta.length === 0) return undefined;
-      const itemId = optionalString(event.item_id);
-      const parsed: ParsedRealtimeServerEvent = {
-        type: REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DELTA,
-        delta: event.delta,
-      };
-      if (itemId) parsed.itemId = itemId;
-      return parsed;
-    }
-    case REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DONE: {
-      const transcript = text(event.transcript);
-      if (!transcript) return undefined;
-      const itemId = optionalString(event.item_id);
-      const parsed: ParsedRealtimeServerEvent = {
-        type: REALTIME_SERVER_EVENT.RESPONSE_OUTPUT_AUDIO_TRANSCRIPT_DONE,
-        transcript,
-      };
-      if (itemId) parsed.itemId = itemId;
-      return parsed;
-    }
-    case REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_DELTA: {
-      // A delta needs its item: the preview it grows is keyed by the turn it
-      // belongs to, and words no turn claims could only be drawn wrong.
-      if (!isWireString(event.delta) || event.delta.length === 0) return undefined;
-      const itemId = text(event.item_id);
-      if (!itemId) return undefined;
-      return {
-        type: REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_DELTA,
-        itemId,
-        delta: event.delta,
-      };
-    }
-    case REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_COMPLETED: {
-      // A transcription that came back empty still ends its turn — whatever
-      // preview its deltas built must leave with it — it just carries
-      // nothing worth a history line, which the recording path refuses on
-      // its own. A failed transcription arrives as its own event, parsed
-      // below, and ends its turn the same way.
-      const itemId = text(event.item_id);
-      if (!itemId) return undefined;
-      return {
-        type: REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_COMPLETED,
-        itemId,
-        transcript: text(event.transcript) ?? "",
-      };
-    }
-    case REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_FAILED: {
-      const itemId = text(event.item_id);
-      if (!itemId) return undefined;
-      return { type: REALTIME_SERVER_EVENT.INPUT_AUDIO_TRANSCRIPTION_FAILED, itemId };
-    }
-    case REALTIME_SERVER_EVENT.INPUT_AUDIO_BUFFER_COMMITTED: {
-      const itemId = text(event.item_id);
-      if (!itemId) return undefined;
-      return { type: REALTIME_SERVER_EVENT.INPUT_AUDIO_BUFFER_COMMITTED, itemId };
-    }
-    case REALTIME_SERVER_EVENT.OUTPUT_AUDIO_BUFFER_STOPPED: {
-      // The drain names the response it drained. An old reply's buffer can
-      // empty after its follow-up was already asked for, and a drain read as
-      // the current reply's would end a turn under audio it never played.
-      const responseId = optionalString(event.response_id);
-      const parsed: ParsedRealtimeServerEvent = {
-        type: REALTIME_SERVER_EVENT.OUTPUT_AUDIO_BUFFER_STOPPED,
-      };
-      if (responseId) parsed.responseId = responseId;
-      return parsed;
-    }
-    case REALTIME_SERVER_EVENT.OUTPUT_AUDIO_BUFFER_STARTED: {
-      // The resume names its response the same way, so a stale start cannot
-      // un-remember a drain that was really the current reply's ending.
-      const responseId = optionalString(event.response_id);
-      const parsed: ParsedRealtimeServerEvent = {
-        type: REALTIME_SERVER_EVENT.OUTPUT_AUDIO_BUFFER_STARTED,
-      };
-      if (responseId) parsed.responseId = responseId;
-      return parsed;
-    }
-    case REALTIME_SERVER_EVENT.RESPONSE_DONE: {
-      const responseId = optionalString(recordField(event, "response")?.id);
-      const hasAudio = audioFromDone(event);
-      const parsed: ParsedRealtimeServerEvent = {
-        type: REALTIME_SERVER_EVENT.RESPONSE_DONE,
-        calls: functionCallsFromDone(event),
-      };
-      if (responseId) parsed.responseId = responseId;
-      if (hasAudio !== undefined) parsed.hasAudio = hasAudio;
-      return parsed;
-    }
-    case REALTIME_SERVER_EVENT.ERROR: {
-      const error = recordField(event, "error");
-      const message = optionalString(error?.message);
-      const eventId = optionalString(error?.event_id);
-      const errorType = optionalString(error?.type);
-      const errorCode = optionalString(error?.code);
-      const parsed: ParsedRealtimeServerEvent = {
-        type: REALTIME_SERVER_EVENT.ERROR,
-        message: message ?? "The voice service reported an error.",
-      };
-      if (eventId) parsed.eventId = eventId;
-      if (errorType) parsed.errorType = errorType;
-      if (errorCode) parsed.errorCode = errorCode;
-      return parsed;
-    }
-    default:
-      return undefined;
-  }
-}
-
-/**
- * Builds the event that asks for the reply voicing a tool's answer. Its tools
- * are withheld: this turn was opened to say what the brain answered, not to ask
- * it again, so an answer that reads like an instruction cannot make it call
- * anything — the same guard every turn Luke opens himself carries.
- */
-export function functionCallFollowUpEvents(): readonly WireRecord[] {
-  return [
-    {
-      type: REALTIME_CLIENT_EVENT.RESPONSE_CREATE,
-      response: {
-        tools: [],
-        tool_choice: "none",
-      },
-    },
-  ];
-}
