+++ packages/settings/package.json
@@ -19,12 +19,12 @@
     "typescript": "catalog:"
   },
   "dependencies": {
+    "@sidecar/actions": "workspace:*",
     "@sidecar/analytics": "workspace:*",
     "@sidecar/calendar": "workspace:*",
     "@sidecar/credentials": "workspace:*",
     "@sidecar/guide": "workspace:*",
     "@sidecar/live": "workspace:*",
-    "@sidecar/realtime": "workspace:*",
     "@sidecar/session": "workspace:*",
     "@sidecar/surface": "workspace:*",
     "@sidecar/wire": "workspace:*"
