+++ packages/settings/src/schema.test.ts
@@ -1,5 +1,6 @@
 import assert from "node:assert/strict";
 import test from "node:test";
+import { isRealtimeVoice } from "@sidecar/actions";
 import { PRODUCT_SETTING_VALUE } from "@sidecar/analytics";
 import { CREDENTIAL_PROVIDER_ID, CREDENTIAL_SOURCE } from "@sidecar/credentials/vocabulary";
 import {
@@ -9,7 +10,6 @@ import {
   isAppSettingId,
 } from "@sidecar/guide";
 import { isLiveVoice, LIVE_DEFAULTS, LIVE_VOICE } from "@sidecar/live";
-import { isRealtimeVoice } from "@sidecar/realtime";
 import { PROVIDER_ID, SUPERSET_WORKSPACE_PROVIDER_ID } from "@sidecar/session";
 import {
   APP_SETTING_SCHEMA,
