+++ tools/ios-parity/package.json
@@ -13,7 +13,6 @@
     "@sidecar/analytics": "workspace:*",
     "@sidecar/guide": "workspace:*",
     "@sidecar/hosted": "workspace:*",
-    "@sidecar/realtime": "workspace:*",
     "@sidecar/session": "workspace:*",
     "@sidecar/wire": "workspace:*"
   },
